Various fixes for test breakage
