+++ tests/backend/unit/services/grid/test_grid_service.py
@@ -23,7 +23,7 @@ class TestGridService(BaseTestCase):
         result = GridService.merge_to_multi_polygon(aoi_geojson, True)
 
         # assert
-        self.assertEqual(str(expected), str(result))
+        self.assertDeepAlmostEqual(expected, result)
 
     def test_feature_collection_to_multi_polygon_nodissolve(self):
         # arrange
@@ -52,7 +52,7 @@ class TestGridService(BaseTestCase):
         result = GridService.trim_grid_to_aoi(grid_dto)
 
         # assert
-        self.assertEqual(str(expected), str(result))
+        self.assertDeepAlmostEqual(expected, result, places=6)
 
     def test_trim_grid_to_aoi_noclip(self):
         # arrange
@@ -67,7 +67,7 @@ class TestGridService(BaseTestCase):
         result = GridService.trim_grid_to_aoi(grid_dto)
 
         # assert
-        self.assertEqual(str(expected), str(result))
+        self.assertDeepAlmostEqual(expected, result)
 
     def test_tasks_from_aoi_features(self):
         # arrange
